Updates performed: - Removed Result and Result<T> classes; - Implemented ExceptionHandlingMiddleware; - Implemented mapping exception type to status code; - Code refactoring; - Fixed some issues found by Sergey Svirsky.

+++ ReadingList.Domain/CommandHandlers/PrivateList/RemovePrivateItemCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using ReadingList.Domain.Commands.PrivateList;
+using ReadingList.Domain.Exceptions;
 using ReadingList.WriteModel;
 
 namespace ReadingList.Domain.CommandHandlers.PrivateList
@@ -16,8 +17,14 @@ namespace ReadingList.Domain.CommandHandlers.PrivateList
 
         protected override async Task Handle(RemovePrivateItemCommand command)
         {
-            var item = await _dbContext.PrivateBookListItems.SingleAsync(i => i.Id == command.Id);
+            var item = await _dbContext.PrivateBookListItems.FirstOrDefaultAsync(i =>
+                i.BookList.Owner.Login == command.UserLogin && i.Id == command.Id);
+            
+            if(item == null)
+                throw new ObjectNotExistException($"Item Id: {command.Id.ToString()}");
+            
             _dbContext.PrivateBookListItems.Remove(item);
+            
             await _dbContext.SaveChangesAsync();
         }
     }
